refactor: extract migration runner and pure function to soundbot.migration
Pure migrate_v1_to_v2 and the on_ready run_migration_if_needed runner
both lived in store.py and bot.py respectively. Splitting them across
two modules forced bot.py to host a top-level non-Cog function whose
only job was to duck-type guild.fetch_soundboard_sounds() — easy to
miss when scanning bot.py for command handlers, and not where
maintainers expect store-format logic to live.

soundbot/migration.py now houses both the pure transformation and the
runner. bot.py imports run_migration_if_needed from there. store.py is
back to being purely about local persistence with no version-bump
plumbing. test_migration.py and test_store.py are updated to import
from the new location; no behavior change.
